✨ refactor(`CaptchaRepository.java`, `CaptchaToken.java`, `CaptchaController.java`, `CaptchaFilter.java`, `CsrfWebFilter.java`): Rename package structure for captcha-related classes; update captcha controller to use DataBuffer for response handling.
