+++ boot/platform/src/main/java/com/plate/boot/security/captcha/CaptchaToken.java
@@ -1,4 +1,4 @@
-package com.plate.boot.security.core.captcha;
+package com.plate.boot.security.captcha;
 
 import org.springframework.util.Assert;
